Feature/access control2

+++ app/controllers/InventoryController.java
@@ -14,13 +14,23 @@ import javax.inject.Inject;
 import java.io.File;
 
 
+import org.pac4j.play.java.UserProfileController;
+import org.pac4j.core.profile.CommonProfile;
+import org.jasig.cas.client.authentication.*;
+import org.pac4j.core.profile.UserProfile;
+import org.pac4j.core.profile.ProfileManager;
+import org.pac4j.play.PlayWebContext;
+import org.pac4j.core.credentials.Credentials;
+import org.pac4j.core.client.Client;
+
 import views.html.inventory.*;
 
 import models.*;
+import models.User;
 
 import org.pac4j.play.java.RequiresAuthentication;
 
-public class InventoryController extends Controller {
+public class InventoryController extends UserProfileController<CommonProfile> {
 
     private final MailerClient mailer;
 
@@ -28,21 +38,38 @@ public class InventoryController extends Controller {
     public InventoryController(MailerClient mailer) {
     this.mailer = mailer;
     }
-	
 
+	@RequiresAuthentication(clientName = "CasClient")
 	@Transactional
-	//@RequiresAuthentication(clientName = "CasClient")
-	public Result index() {
+	public Result index(){
+		if(!checkPrivileges())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/home");
+		}
+		
 		return ok(index.render());
 	}
 
 	@Transactional
 	public Result show(Integer id) {
+		if(!checkPrivilegesAdmin())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/items");
+		}
+		
 		return ok(show.render(Inventory.findById(id)));
 	}
 
 	@Transactional(readOnly=true)
 	public Result edit(Integer id) {
+		if(!checkPrivilegesAdmin())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/items");
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class).fill(
 			Inventory.findById(id));
 		return ok(edit.render(id, invForm));
@@ -50,11 +77,23 @@ public class InventoryController extends Controller {
 
 	@Transactional(readOnly=true)
 	public Result create() {
+		if(!checkPrivilegesAdmin())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/items");
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class);
 		return ok(create.render(invForm));
 	}
 	@Transactional
 	public Result save() {
+		if(!checkPrivileges())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/items");
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class).bindFromRequest();
 
 		if(invForm.hasErrors()) {
@@ -66,12 +105,24 @@ public class InventoryController extends Controller {
 	}
 	@Transactional
 	public Result delete(Integer id) {
+		if(!checkPrivilegesAdmin())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/items");
+		}
+		
 		Inventory.findById(id).delete();
 		flash("success", "Inventory item has been deleted");
 		return ok(index.render());
 	}
 	@Transactional
 	public Result update(Integer id) {
+		if(!checkPrivileges())
+		{
+			flash("error", "Insufficient Privileges");
+			return redirect("/items");
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class).bindFromRequest();
 		if(invForm.hasErrors()) {
 			return badRequest(edit.render(id, invForm));
@@ -80,5 +131,28 @@ public class InventoryController extends Controller {
 			flash("success", "Item " + id + " has been updated");
 			return ok(show.render(Inventory.findById(id)));
 		}		
+	}
+	
+	//This function returns false if current user does not possess any roles
+	public boolean checkPrivileges()
+	{
+		CommonProfile profile = getUserProfile();
+		//if you don't have admin role then redirect back to dashboard
+		if((User.findById(profile.getId()).role == null))
+		{
+			return false;
+		}
+		return true;
+	}
+	//This function returns false if current user does not possess the role Admin or SuperAdmin
+	public boolean checkPrivilegesAdmin()
+	{
+		CommonProfile profile = getUserProfile();
+		//if you don't have admin role then redirect back to dashboard
+		if(User.findById(profile.getId()).role == User.Role.Admin ||  User.findById(profile.getId()).role == User.Role.SuperAdmin)
+		{
+			return true;
+		}
+		return false;
 	}
 }
